requirements.txt
streamlit==1.31.1
google-generativeai==0.3.2
python-dotenv==1.0.1 
openai==1.65.4
